% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{clr.epsilon}
\alias{clr.epsilon}
\title{Modified centered log-ratio with epsilon}
\usage{
clr.epsilon(mat, base = exp(1), const = NULL)
}
\arguments{
\item{mat}{Numeric matrix.}

\item{base}{Numeric; log base.}

\item{const}{Numeric; constant to add.}
}
\value{
Transformed matrix.
}
\description{
Modified centered log-ratio with epsilon
}
\details{
The usual CLR transformation adds a pseudocount to all zeros or all counts, causing artificial inflation of negative values. By contrast, the modified CLR transformation transforms the positive counts and ensures that the original positive counts are all greater than zero.
}
