% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{clr}
\alias{clr}
\title{Centered log-ratio transformation}
\usage{
clr(x, base = exp(1), use.pseudo = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{base}{Numeric; log base.}

\item{use.pseudo}{Logical; add 1 for zero handling?}
}
\value{
Transformed vector.
}
\description{
Centered log-ratio transformation
}
