% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{filter_by_min_abundance_min_sample}
\alias{filter_by_min_abundance_min_sample}
\title{Function to filter taxa by relative abundances and prevalence}
\usage{
filter_by_min_abundance_min_sample(otu_table, min_abundance, min_sample)
}
\arguments{
\item{otu_table}{a matrix with rows being samples}

\item{min_abundance}{threshold for minimum relative abundances}

\item{min_sample}{threshold for minimum prevalence}
}
\value{
A matrix of relative abundances containing the filtered features
}
\description{
Function to filter taxa by relative abundances and prevalence
}
\details{
This function takes an OTU count matrix and returns a matrix of relative abundances with
only features that pass the threshold.
}
