% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gm.mean.pos}
\alias{gm.mean.pos}
\title{Geometric mean of positive components only}
\usage{
gm.mean.pos(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Geometric mean of positive entries.
}
\description{
Geometric mean of positive components only
}
