% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gm.mean}
\alias{gm.mean}
\title{Geometric mean of positive values}
\usage{
gm.mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Logical; remove NA values?}
}
\value{
Geometric mean.
}
\description{
Computes the geometric mean of positive and non-NA entries.
}
