% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{GowerCentering}
\alias{GowerCentering}
\title{Double center a matrix using Gower centering}
\usage{
GowerCentering(M)
}
\arguments{
\item{M}{Numeric matrix of dimension n x n.}
}
\value{
A double-centered matrix.
}
\description{
Double center a matrix using Gower centering
}
