% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{procrustes}
\alias{procrustes}
\title{Procrustes analysis between two matrices}
\usage{
procrustes(A, B)
}
\arguments{
\item{A}{Numeric matrix.}

\item{B}{Numeric matrix.}
}
\value{
List with normalized matrices, rotation, transformed B, and RSS.
}
\description{
Procrustes analysis between two matrices
}
