% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rclr}
\alias{rclr}
\title{Robust centered log-ratio transform}
\usage{
rclr(x, base = exp(1))
}
\arguments{
\item{x}{Numeric vector.}

\item{base}{Numeric; log base.}
}
\value{
Transformed vector.
}
\description{
Robust centered log-ratio transform
}
\details{
This function transform nonzero counts using the centered log ratio while leaving zero counts unmodified.
}
