% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rlog}
\alias{rlog}
\title{Log transform positive counts}
\usage{
rlog(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector with log-transformed non-zero entries.
}
\description{
Replaces non-zero entries with their natural log.
}
